requirements.txt
appdirs==1.4.3
backports.ssl-match-hostname==3.5.0.1
click==6.7
Flask==0.12
Flask-SocketIO==2.8.6
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
pyserial==3.3
python-engineio==1.3.1
python-socketio==1.7.3
requests==2.13.0
six==1.10.0
socketIO-client==0.7.2
websocket-client==0.40.0
Werkzeug==0.12.1
